requirements.txt
streamlit
google-generativeai
cloudinary
gspread
oauth2client
pandas
Pillow